@@ -15,21 +15,21 @@
 
 name: example
 description: Sample of using instant_xml_parser to generate a parser via annotation.
-version: 0.9.0
+version: 1.0.0-nullsafety.0
 homepage: https://github.com/sarahec/instant_xml_parser
 
 environment:
-  sdk: '>=2.10.0 <3.0.0'
+  sdk: '>=2.12.0-0 <3.0.0'
 
 dependencies:
   logging: ^1.0.0-nullsafety.0
   # xml: ^5.0.0-nullsafety.1
-  ixp_runtime: ^0.9.0
+  ixp_runtime: ^1.0.0-nullsafety.0
 
 dev_dependencies:
   build_runner: ^1.0.0
   build_test: ^0.10.3
-  test: ^1.14.4
+  test: ^1.16.0-nullsafety.13
   instant_xml_parser: ^0.9.0
 
 dependency_overrides:
@@ -37,5 +37,6 @@ dependency_overrides:
     path: ../instant_xml_parser  
   ixp_runtime:
     path: ../ixp_runtime
+  logging: ^1.0.0-nullsafety.0
   coverage:
     path: ../../coverage
